Update class-reflection-extensions.md
All types mentioned in the examples except `Type` come from `PHPStan\Reflection`, which means they don't need to be imported. I personally tripped over Type not being in the place where I expected it to be. This PR aims to make the copy, paste, implement cycle a bit easier for the developer.